vueで作った画面を立ち上げるjsとvueが関係ないロード時のjsを分けた（それぞれviews.js, modules.js） #297
