+++ app/javascript/views.js
@@ -6,14 +6,6 @@ import MarksEdit from './views/marks_edit.vue'
 import ReportsIndex from './views/reports_index.vue'
 import tags from './modules/tags'
 
-window.addEventListener('popstate', () => {
-  tags.browserBack()
-})
-tags.homeTop()
-const homeIndex = document.getElementById('home-index')
-const helpIndex = document.getElementById('help-index')
-if (homeIndex || helpIndex) tags.displayHome()
-
 document.addEventListener('turbo:load', () => {
   const selectors = [
     '#rooms-show',
